Change accession in amplicon test private study to avoid clash
